+++ database/database.go
@@ -1,7 +1,9 @@
 package database
 
 import (
-	"github.com/lib/pq"
+	"time"
+
+	"github.com/pikapika/models"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 )
@@ -20,34 +22,22 @@ func SetDB(database *gorm.DB) {
 	db = database
 }
 
-type Event struct {
-	gorm.Model
-	Message            []byte         `json:"message"`
-	Source             string         `json:"source"`
-	DestinationAddress string         `json:"destinationAddress"`
-	Topics             pq.StringArray `gorm:"type:text[]" json:"topics"`
-	NotificationType   string         `json:"notificationtype"`
-	IdempotencyKey     string         `json:"idempotencykey"`
-	Status             string         `json:"status"`
-	Attempts           int            `json:"attempts"`
-	NextRetry          int64          `json:"nextRetry"`
-}
-
-func CreateEvent(event *Event) error {
+func CreateEvent(event *models.Event) error {
 	return db.Create(&event).Error
 }
 
-func UpdateEvent(event *Event) error {
+func UpdateEvent(event *models.Event) error {
 	return db.Model(&event).Update("Status", event.Status).Error
 }
 
-func GetIncompleteEvents() ([]Event, error) {
-	var events []Event
-	err := db.Where("status = ?", "Not Completed Yet").Find(&events).Error
+func GetEventsForRetry() ([]*models.Event, error) {
+	var events []*models.Event
+	err := db.Where("status = ? AND NextRetry > ?", "Not Completed Yet", time.Now().Unix()).Find(&events).Error
 	return events, err
 }
+
 func Initialize() error {
-	err := db.AutoMigrate(&Event{})
+	err := db.AutoMigrate(&models.Event{})
 	if err != nil {
 		return err
 	}
