+++ go.work
@@ -7,4 +7,5 @@ use (
 	./services/notification_services
 	./database
 	./Timer
+	./models
 )
