+++ models/go.mod
@@ -1,0 +1,3 @@
+module github.com/pikapika/models
+
+go 1.22.0
